refactor(vm): split skia rendering sources
