+++ .agent/evidence/skia-generated-image.jsonl
@@ -2,3 +2,7 @@
 {"copyright":"Copyright (C) 2026 Amalgam Solucoes em TI Ltda","spdx":"LGPL-2.1-only","date":"2026-07-23T02:44:00-03:00","revision":"894febba7dd1c188b91f2af938028b5274126a19","milestone":"0","target":"source-inspection","command":"unzip -q /tmp/Tcsort.zip -d /tmp/tcsort-inspect.CD298m; nl -ba /tmp/tcsort-inspect.CD298m/tcsort.java; aapt dump badging /tmp/tcsort-inspect.CD298m/Tcsort.apk","status":"pass","entryPoint":"Tcsort.initUI -> testGraphics","imageType":"totalcross.io.device.printer.MonoImage","dimensions":"576x576","drawing":"fillRect(0, y, 576, 576) with Color.WHITE; drawRect(10, y + 10, 556, 556) with Color.BLACK; inherited Control.y is zero for this new MainWindow","output":"Settings.appPath/nome.png via Image.createPng(File)","launch":"Vm.exec(\"viewer\", path, 0, true)","package":"totalcross.appphdb","activity":"totalcross.appphdb.Loader","expectedPixels":{"background":"opaque white","border":"opaque black at inset 10","interior":"opaque white"}}
 {"copyright":"Copyright (C) 2026 Amalgam Solucoes em TI Ltda","spdx":"LGPL-2.1-only","date":"2026-07-23T02:44:00-03:00","revision":"894febba7dd1c188b91f2af938028b5274126a19","milestone":"0","target":"android-baseline","command":"adb install -r /tmp/tcsort-inspect.CD298m/Tcsort.apk; adb shell am start -n totalcross.appphdb/totalcross.appphdb.Loader; sleep 8; adb pull /data/user/0/totalcross.appphdb/nome.png /tmp/tcsort-baseline-nome.png; ffmpeg -v error -i /tmp/tcsort-baseline-nome.png -f rawvideo -pix_fmt rgba /tmp/tcsort-baseline-nome.rgba","status":"expected-failure","device":"emulator-5554","assertions":1,"assertion":"decoded RGBA stream has zero non-zero bytes","observed":"PNG 576x576 RGBA; nonzero_bytes=0; output is fully transparent/blank","artifacts":["/tmp/tcsort-baseline-nome.png","/tmp/tcsort-baseline-nome.rgba","/tmp/tcsort-baseline-logcat.log","/tmp/tcsort-baseline-screen.png"],"artifactHashes":{"png":"f73f4977d1a7f2640681b824dc820c657a33886ab941d6d567cd363d7c4d631","screen":"1d1c2a10355f735782a3766b824d89cc9562468f0b822301468cd01ec6439f20"},"limitation":"The baseline was reproduced on the connected emulator, not on a physical Android device."}
 {"copyright":"Copyright (C) 2026 Amalgam Solucoes em TI Ltda","spdx":"LGPL-2.1-only","date":"2026-07-23T02:55:00-03:00","revision":"9bf9faeb31eaa72933676ab96fe6f38bdd7581d2","milestone":"1","target":"source-relocation","command":"git mv TotalCrossVM/src/nm/ui/android/skia.h TotalCrossVM/src/nm/ui/skia/skia.h; git mv TotalCrossVM/src/nm/ui/android/skia.cpp TotalCrossVM/src/nm/ui/skia/skia.cpp; update CMake and native include paths; git diff --check","status":"pass","changedPaths":["TotalCrossVM/CMakeLists.txt","TotalCrossVM/src/nm/ui/GraphicsPrimitives.h","TotalCrossVM/src/nm/ui/PalmFont_c.h","TotalCrossVM/src/nm/ui/android/gfx_Graphics_c.h","TotalCrossVM/src/nm/ui/font_Font.c","TotalCrossVM/src/nm/ui/image_Image.c","TotalCrossVM/src/nm/ui/linux/gfx_Graphics_c.h","TotalCrossVM/src/nm/ui/skia/skia.h","TotalCrossVM/src/nm/ui/skia/skia.cpp"],"assertions":2,"assertionsDetail":["CMake references nm/ui/skia/skia.cpp","tracked native includes reference skia/skia.h or the correct relative path"],"limitation":"The structural CMake/Ninja build is deferred until the remaining source split is complete; the ignored generated tcvm.xcodeproj was not changed."}
+{"copyright":"Copyright (C) 2026 Amalgam Solucoes em TI Ltda","spdx":"LGPL-2.1-only","date":"2026-07-23T03:03:18-03:00","revision":"WORKTREE_AFTER_24ca5c87194d309f017de0b469d14a2923ad2165","milestone":"1","target":"source-split","command":"update CMake with skia_surface.cpp and skia_primitives.cpp; extract skia_internal.h, skia_surface.cpp, skia_primitives.cpp and GraphicsPrimitives{Skia,Text,Shapes,Screen}_c.h; preserve original include positions; git diff --check","status":"pass","changedPaths":["TotalCrossVM/src/nm/ui/skia/skia_internal.h","TotalCrossVM/src/nm/ui/skia/skia_surface.cpp","TotalCrossVM/src/nm/ui/skia/skia_primitives.cpp","TotalCrossVM/src/nm/ui/GraphicsPrimitivesSkia_c.h","TotalCrossVM/src/nm/ui/GraphicsPrimitivesText_c.h","TotalCrossVM/src/nm/ui/GraphicsPrimitivesShapes_c.h","TotalCrossVM/src/nm/ui/GraphicsPrimitivesScreen_c.h","TotalCrossVM/src/nm/ui/GraphicsPrimitives_c.h"],"assertions":3,"assertionsDetail":["no skia source remains under TotalCrossVM/src/nm/ui/android","skia.cpp is 350 lines and surface/primitives units are 167/255 lines","CMake lists all three Skia translation units"],"limitation":"The fragments are included at their original positions to preserve C preprocessor order; no behavior change was intended."}
+{"copyright":"Copyright (C) 2026 Amalgam Solucoes em TI Ltda","spdx":"LGPL-2.1-only","date":"2026-07-23T03:03:18-03:00","revision":"WORKTREE_AFTER_24ca5c87194d309f017de0b469d14a2923ad2165","milestone":"1","target":"native-structure-build","command":"cmake -S TotalCrossVM -B build-skia-structure -DCMAKE_BUILD_TYPE=Release -G Ninja; ninja -C build-skia-structure","status":"pass","assertions":2,"observed":"CMake configured successfully and Ninja linked build-skia-structure/libtcvm.dylib; final Ninja run completed with only duplicate-library linker warnings","logs":["/tmp/skia-m1-cmake.log","/tmp/skia-m1-ninja-rerun4.log"],"limitation":"This is a macOS arm64 structural build; it does not validate Android deployment or bitmap authority."}
+{"copyright":"Copyright (C) 2026 Amalgam Solucoes em TI Ltda","spdx":"LGPL-2.1-only","date":"2026-07-23T03:03:18-03:00","revision":"WORKTREE_AFTER_24ca5c87194d309f017de0b469d14a2923ad2165","milestone":"1","target":"sdk-build","command":"cd TotalCrossSDK && ./gradlew-agent clean dist --warning-mode=none --console=plain","status":"pass","assertions":1,"observed":"BUILD SUCCESSFUL in 1m 5s; 22 actionable tasks, 20 executed, 2 up-to-date, 0 failed tasks","log":"TotalCrossSDK/agent-logs/20260723-030119-clean-agent.log","limitation":"The build emitted existing Java compilation and Javadoc warnings summarized by Gradle Agent."}
+{"copyright":"Copyright (C) 2026 Amalgam Solucoes em TI Ltda","spdx":"LGPL-2.1-only","date":"2026-07-23T03:03:18-03:00","revision":"WORKTREE_AFTER_24ca5c87194d309f017de0b469d14a2923ad2165","milestone":"1","target":"registered-tests","command":"ctest --test-dir build-skia-structure --output-on-failure","status":"pass-no-tests","assertions":0,"observed":"CTest returned 0 and reported No tests were found!!!","limitation":"The repository does not register focused graphics tests in this structural CMake build."}
