+++ .agent/state/skia-generated-image.md
@@ -6,19 +6,28 @@ SPDX-License-Identifier: LGPL-2.1-only
 
 # Estado do ExecPlan Skia Generated Image
 
-Milestone ativo: nenhum; Milestone 0 concluído. Próximo milestone: 1, mover e
-dividir as fontes sem alteração intencional de comportamento.
+Milestone ativo: nenhum; Milestone 1 concluído. Próximo milestone: 2, adicionar
+superfícies Skia próprias das imagens e encaminhar o desenho para o canvas
+selecionado.
 
-Último commit lógico: `9bf9faeb3 refactor(vm): relocate skia sources`.
+Último commit lógico: `refactor(vm): split skia rendering sources` (commit
+atual).
 
-Caminhos alterados neste milestone até agora: `TotalCrossVM/CMakeLists.txt`,
+Caminhos alterados neste milestone: `TotalCrossVM/CMakeLists.txt`,
 `TotalCrossVM/src/nm/ui/GraphicsPrimitives.h`,
+`TotalCrossVM/src/nm/ui/GraphicsPrimitives_c.h`,
+`TotalCrossVM/src/nm/ui/GraphicsPrimitivesSkia_c.h`,
+`TotalCrossVM/src/nm/ui/GraphicsPrimitivesText_c.h`,
+`TotalCrossVM/src/nm/ui/GraphicsPrimitivesShapes_c.h`,
+`TotalCrossVM/src/nm/ui/GraphicsPrimitivesScreen_c.h`,
 `TotalCrossVM/src/nm/ui/PalmFont_c.h`,
 `TotalCrossVM/src/nm/ui/android/gfx_Graphics_c.h`,
 `TotalCrossVM/src/nm/ui/font_Font.c`,
 `TotalCrossVM/src/nm/ui/image_Image.c`,
 `TotalCrossVM/src/nm/ui/linux/gfx_Graphics_c.h`, e os arquivos relocados
 `TotalCrossVM/src/nm/ui/skia/skia.h` e `skia.cpp`.
+Também foram criados `skia_internal.h`, `skia_surface.cpp` e
+`skia_primitives.cpp`.
 
 Anexo inspecionado: `Tcsort.zip`, baixado de
 `https://github.com/user-attachments/files/30019877/Tcsort.zip`, SHA-256
@@ -35,16 +44,19 @@ RGBA decodificado teve `nonzero_bytes=0`, confirmando a saída transparente e
 vazia. A captura de tela, o logcat e o arquivo gerado permanecem em `/tmp`;
 os caminhos e hashes estão no JSONL de evidência.
 
-Validação executada: download/listagem/extração do anexo, inspeção da fonte,
-reprodução baseline no Android e `git diff --check` antes do commit de
-relocação. O build estrutural do Milestone 1 ainda não foi executado; nenhum
-teste ou validação de milestone posterior foi executado.
+Validação do milestone: `cmake -S TotalCrossVM -B build-skia-structure
+-DCMAKE_BUILD_TYPE=Release -G Ninja` passou; `ninja -C build-skia-structure`
+passou e gerou `libtcvm.dylib`; em `TotalCrossSDK`,
+`./gradlew-agent clean dist --warning-mode=none --console=plain` passou com 0
+ tarefas falhas; `ctest --test-dir build-skia-structure --output-on-failure`
+retornou sucesso, mas sem testes registrados. Logs do CMake/Ninja estão em
+`/tmp/skia-m1-*.log`; o resumo completo do SDK está em
+`TotalCrossSDK/agent-logs/20260723-030119-clean-agent.log`.
 
 Bloqueios: nenhum bloqueio técnico para o Milestone 1. O arquivo de estado
 solicitado no prompt e o nome originalmente citado no plano estavam ausentes;
 este arquivo segue o nome canônico definido na seção “Working Set and Resume
 Protocol” do ExecPlan.
 
-Próxima ação: dividir `skia.cpp` em unidades lógicas e separar os blocos de
-`GraphicsPrimitives_c.h`, depois executar somente o build estrutural previsto
-para fechar o Milestone 1.
+Próxima ação: iniciar somente quando solicitado o Milestone 2, adicionando
+superfícies Skia próprias das imagens; não executar ainda suas validações.
