+++ TotalCrossVM/src/nm/ui/GraphicsPrimitivesScreen_c.h
@@ -1,0 +1,1 @@
+// Copyright (C) 2026 Amalgam Solucoes em TI Ltda//// SPDX-License-Identifier: LGPL-2.1-only#ifndef GRAPHICSPRIMITIVESSCREEN_C_H#define GRAPHICSPRIMITIVESSCREEN_C_Hstatic bool startupGraphics(int16 appTczAttr) // there are no threads running at this point{    return graphicsStartup(&screen, appTczAttr);}static bool createScreenSurface(Context currentContext, bool isScreenChange){   bool ret = false;   if (screen.screenW <= 0 || screen.screenH <= 0)      return false;   if (graphicsCreateScreenSurface(&screen))   {      TCObject *screenObj;      screenObj = getStaticFieldObject(currentContext,loadClass(currentContext, "totalcross.ui.gfx.Graphics",false), "mainWindowPixels");#ifdef darwin // in darwin, the pixels buffer is pre-initialized and never changed      if (controlEnableUpdateScreenPtr == null)         controlEnableUpdateScreenPtr = getStaticFieldInt(loadClass(currentContext, "totalcross.ui.Control",false), "enableUpdateScreen");      ret = true;#else      if (isScreenChange)      {         screen.mainWindowPixels = *screenObj = null;         gc(currentContext); // let the gc collect the old screen object      }      else      {         controlEnableUpdateScreenPtr = getStaticFieldInt(loadClass(currentContext, "totalcross.ui.Control",false), "enableUpdateScreen");      }      *screenObj = screen.mainWindowPixels = createArrayObject(currentContext, INT_ARRAY, screen.screenW * screen.screenH);      setObjectLock(*screenObj, UNLOCKED);      ret = screen.mainWindowPixels != null && controlEnableUpdateScreenPtr != null;#endif   }   return ret;}void markWholeScreenDirty(Context currentContext){   LOCKVAR(screen);   currentContext->dirtyX1 = currentContext->dirtyY1 = 0;   currentContext->dirtyX2 = screen.screenW;   currentContext->dirtyY2 = screen.screenH;   currentContext->fullDirty = true;   UNLOCKVAR(screen);}static bool checkScreenPixels(){   return screen.pixels != null;}void updateScreen(Context currentContext){#ifdef ANDROID   if (appPaused) return;#endif   LOCKVAR(screen);   if (keepRunning && checkScreenPixels() && controlEnableUpdateScreenPtr && *controlEnableUpdateScreenPtr && (currentContext->fullDirty || (currentContext->dirtyX1 != screen.screenW && currentContext->dirtyX2 != 0 && currentContext->dirtyY1 != screen.screenH && currentContext->dirtyY2 != 0)))   {      if (updateScreenBits(currentContext)) // move the temporary buffer to the real screen      {#ifdef darwin         currentContext->dirtyX1 = screen.screenW;         currentContext->dirtyY1 = screen.screenH;         currentContext->dirtyX2 = currentContext->dirtyY2 = 0;         currentContext->fullDirty = false;         UNLOCKVAR(screen); // without this, a deadlock can occur in iOS if the user minimizes the application, since another thread can trigger a markScreenDirty#endif         graphicsUpdateScreen(currentContext, &screen);#ifdef darwin         LOCKVAR(screen);#else         currentContext->dirtyX1 = screen.screenW;         currentContext->dirtyY1 = screen.screenH;         currentContext->dirtyX2 = currentContext->dirtyY2 = 0;         currentContext->fullDirty = false;#endif      }   }   UNLOCKVAR(screen);#if defined ANDROID || defined darwin   if (desiredScreenShiftY != UNDEFINED_SHIFTY)   {      setShiftYgl(desiredScreenShiftY);      desiredScreenShiftY = UNDEFINED_SHIFTY;   }#endif}void graphicsDestroyPrimitives(){   xfree(lookupR);   xfree(lookupG);   xfree(lookupB);   xfree(lookupGray);   fontDestroy();}/////////////// End of Device-dependant functions ///#endif
