+++ TotalCrossVM/src/nm/ui/GraphicsPrimitives_c.h
@@ -220,436 +220,7 @@ void markDirty(Context currentContext, TCObject surface, int x, int y, int w, in
 
 // This is the main routine that draws a surface (a Control or an Image) in the destination GfxSurface.
 // Destination is always a Graphics object.
-#ifndef SKIA_H
-static void drawSurface(Context currentContext, TCObject dstSurf, TCObject srcSurf, int32 srcX, int32 srcY, int32 width, int32 height,
-                       int32 dstX, int32 dstY, int32 doClip)
-{
-   uint32 i;
-   Pixel * srcPixels;
-   Pixel * dstPixels;
-   int32 srcPitch, srcWidth, srcHeight, alphaMask = 0;
-   bool isSrcScreen = !Surface_isImage(srcSurf);
-   bool unlockSrc = false;
-   if (Surface_isImage(srcSurf))
-   {                                
-#ifdef __gl2_h_ // for opengl, we will use the smoothScaled only if we will draw on an image. for win32, we will always use smoothScale
-      bool forcedSmoothScale = Graphics_isImageSurface(dstSurf); // the destination is always a Graphics object
-#else
-      bool forcedSmoothScale = true;
-#endif
-      srcPitch = srcWidth = (int32)(Image_width(srcSurf) * Image_hwScaleW(srcSurf));
-      srcHeight = (int32)(Image_height(srcSurf) * Image_hwScaleH(srcSurf));
-      if (forcedSmoothScale && (Image_hwScaleW(srcSurf) != 1 || Image_hwScaleH(srcSurf) != 1))
-      {
-         static Method mGetScaledInstance, mGetSmoothScaledInstance;
-         TCObject newSurf;
-         if (mGetScaledInstance == null)
-         {
-            mGetScaledInstance       = getMethod(OBJ_CLASS(srcSurf), false, "getScaledInstance", 2, J_INT, J_INT);
-            mGetSmoothScaledInstance = getMethod(OBJ_CLASS(srcSurf), false, "getSmoothScaledInstance", 2, J_INT, J_INT);
-         }
-         disableGC = true; // the gc may collect the image before we lock it here (*)
-         newSurf = executeMethod(currentContext, Image_hwScaleW(srcSurf) < 1 && Image_hwScaleH(srcSurf) < 1 ? mGetSmoothScaledInstance : mGetScaledInstance, srcSurf, srcWidth, srcHeight).asObj;
-         if (newSurf == null || newSurf == (TCObject)0xFFFFFFFF)
-         {
-            currentContext->thrownException = null;
-            return;
-         }
-         else
-         {
-            srcSurf = newSurf;
-            setObjectLock(newSurf, LOCKED); // (*)
-            unlockSrc = true;
-         }
-         disableGC = false;
-         srcPitch = srcWidth = Image_width(srcSurf);
-         srcHeight = Image_height(srcSurf);
-      }
-      alphaMask = Image_alphaMask(srcSurf);
-   }
-   else
-   {
-      srcPitch = srcWidth = screen.screenW;
-      srcHeight = screen.screenH;
-   }
-   dstPixels = getSurfacePixels(dstSurf);
-   srcPixels = getSurfacePixels(srcSurf);
-   if (!doClip)
-   {
-      /*
-      | Even if no clip is required, we still have to make sure that the
-      | area of the bitmap that we want to copy is inside its area.
-      */
-      if (srcX <= -width  || srcX >= srcWidth || srcY <= -height || srcY >= srcHeight)
-         goto end;
-      dstX += Graphics_transX(dstSurf);
-      dstY += Graphics_transY(dstSurf);
-   }
-   else
-   {
-      dstX += Graphics_transX(dstSurf);
-      dstY += Graphics_transY(dstSurf);
-
-      /* clip the source rectangle to the source surface */
-      if (srcX < 0)
-      {
-         width += srcX;
-         dstX -= srcX;
-         srcX = 0;
-      }
-      i = srcWidth - srcX;
-      if (width > (int32)i)
-         width = i;
-      if (srcY < 0)
-      {
-         height += srcY;
-         dstY -= srcY;
-         srcY = 0;
-      }
-      i = srcHeight - srcY;
-      if (height > (int32)i)
-         height = i;
-
-      /* clip the destination rectangle against the clip rectangle */
-      if (dstX < Graphics_clipX1(dstSurf))
-      {
-         i = Graphics_clipX1(dstSurf) - dstX;
-         dstX = Graphics_clipX1(dstSurf);
-         srcX += i;
-         width -= i;
-      }
-      if ((dstX + width) > Graphics_clipX2(dstSurf))
-         width = Graphics_clipX2(dstSurf) - dstX;
-      if (dstY < Graphics_clipY1(dstSurf))
-      {
-         i = Graphics_clipY1(dstSurf) - dstY;
-         dstY = Graphics_clipY1(dstSurf);
-         srcY += i;
-         height -= i;
-      }
-      if ((dstY + height) > Graphics_clipY2(dstSurf))
-         height = Graphics_clipY2(dstSurf) - dstY;
-
-      /* check the validity */
-      if (width <= 0 || height <= 0)
-         goto end;
-   }
-
-   srcPixels += srcY * srcPitch + srcX;
-   dstPixels += dstY * Graphics_pitch(dstSurf) + dstX;
-#ifdef __gl2_h_
-   if (isSrcScreen)
-      glGetPixels(dstPixels,srcX,srcY,width,height,Graphics_pitch(dstSurf));
-   else
-   if (Graphics_useOpenGL(dstSurf))
-   {
-      int32 fc;
-      int frame;
-
-      if (Image_changed(srcSurf))
-         applyChanges(currentContext, srcSurf);
-      fc = Image_frameCount(srcSurf);
-      frame = (fc <= 1) ? 0 : Image_currentFrame(srcSurf);
-      Image_lastAccess(srcSurf) = getTimeStamp();
-/*img*/ glDrawTexture(Image_textureId(srcSurf), srcX+frame*srcPitch,srcY,width,height, dstX,dstY, 0,0, fc > 1 ? (int32)(Image_widthOfAllFrames(srcSurf) * Image_hwScaleW(srcSurf)) : srcWidth,srcHeight, null, alphaMask);
-   }
-   else
-#endif
-   for (i=0; i < (uint32)height; i++) // in opengl, only case of image drawing on image
-   {
-      PixelConv *ps = (PixelConv*)srcPixels;
-      PixelConv *pt = (PixelConv*)dstPixels;
-      uint32 count = width;
-      if (isSrcScreen)
-         for (;count != 0; pt++,ps++, count--)
-         {
-            pt->pixel = ps->pixel;
-            pt->a = 0xFF;
-         }
-      else
-         for (;count != 0; pt++,ps++, count--)
-         {
-            int32 a = ps->a * alphaMask;
-            a = (a+1 + (a >> 8)) >> 8; // alphaMask * a / 255
-            if (a == 0xFF)
-               pt->pixel = ps->pixel;
-            else
-            if (a != 0)
-            {
-               int32 ma = 0xFF-a;
-               int32 r = (a * ps->r + ma * pt->r);
-               int32 g = (a * ps->g + ma * pt->g);
-               int32 b = (a * ps->b + ma * pt->b);
-               pt->r = (r+1 + (r >> 8)) >> 8; // fast way to divide by 255
-               pt->g = (g+1 + (g >> 8)) >> 8;
-               pt->b = (b+1 + (b >> 8)) >> 8;
-            }
-         }
-      srcPixels += srcPitch;
-      dstPixels += Graphics_pitch(dstSurf);
-   }
-#ifndef __gl2_h_
-   if (!currentContext->fullDirty && !Graphics_isImageSurface(dstSurf)) markScreenDirty(currentContext, dstX, dstY, width, height);
-#else
-   if (Graphics_isImageSurface(dstSurf))
-      Image_changed(Graphics_surface(dstSurf)) = true;
-   else
-      currentContext->fullDirty = true;
-#endif
-end:
-   if (unlockSrc)
-      setObjectLock(srcSurf, UNLOCKED);
-}
-#else
-static void drawSurface(Context currentContext, TCObject dstSurf, TCObject srcSurf, int32 srcX, int32 srcY, int32 w, int32 h,
-   int32 dstX, int32 dstY, int32 doClip) {
-   if (Surface_isImage(srcSurf)) {
-      int32 srcWidth = (int32)(Image_width(srcSurf) * Image_hwScaleW(srcSurf));
-      int32 srcHeight = (int32)(Image_height(srcSurf) * Image_hwScaleH(srcSurf));
-      double scaleW = Image_hwScaleW(srcSurf);
-      double scaleH = Image_hwScaleH(srcSurf);
-      int32 frameCount = Image_frameCount(srcSurf);
-      int32 frame = 0;
-      bool clipSet = false;
-
-      if (scaleW <= 0 || scaleH <= 0 || w <= 0 || h <= 0 || srcWidth <= 0 || srcHeight <= 0) {
-         return;
-      }
-
-      if (!doClip) {
-         if (srcX <= -w || srcX >= srcWidth || srcY <= -h || srcY >= srcHeight) {
-            return;
-         }
-         dstX += Graphics_transX(dstSurf);
-         dstY += Graphics_transY(dstSurf);
-      }
-      else {
-         int32 i;
-
-         dstX += Graphics_transX(dstSurf);
-         dstY += Graphics_transY(dstSurf);
-
-         if (srcX < 0) {
-            w += srcX;
-            dstX -= srcX;
-            srcX = 0;
-         }
-         i = srcWidth - srcX;
-         if (w > i) {
-            w = i;
-         }
-         if (srcY < 0) {
-            h += srcY;
-            dstY -= srcY;
-            srcY = 0;
-         }
-         i = srcHeight - srcY;
-         if (h > i) {
-            h = i;
-         }
-
-         if (dstX < Graphics_clipX1(dstSurf)) {
-            i = Graphics_clipX1(dstSurf) - dstX;
-            dstX = Graphics_clipX1(dstSurf);
-            srcX += i;
-            w -= i;
-         }
-         if ((dstX + w) > Graphics_clipX2(dstSurf)) {
-            w = Graphics_clipX2(dstSurf) - dstX;
-         }
-         if (dstY < Graphics_clipY1(dstSurf)) {
-            i = Graphics_clipY1(dstSurf) - dstY;
-            dstY = Graphics_clipY1(dstSurf);
-            srcY += i;
-            h -= i;
-         }
-         if ((dstY + h) > Graphics_clipY2(dstSurf)) {
-            h = Graphics_clipY2(dstSurf) - dstY;
-         }
-
-         if (w <= 0 || h <= 0) {
-            return;
-         }
-      }
-
-      if (Graphics_isImageSurface(dstSurf)) {
-         TCObject pixelsObj = frameCount > 1 ? Image_pixelsOfAllFrames(srcSurf) : Image_pixels(srcSurf);
-         Pixel *srcPixels = (Pixel*)ARRAYOBJ_START(pixelsObj);
-         Pixel *dstPixels = getSurfacePixels(dstSurf);
-         int32 srcPitch = frameCount > 1 ? Image_widthOfAllFrames(srcSurf) : Image_width(srcSurf);
-         int32 dstPitch = Graphics_pitch(dstSurf);
-         int32 dstRow, row, col;
-         int32 alphaMask = Image_alphaMask(srcSurf);
-
-         if (frameCount > 1) {
-            frame = Image_currentFrame(srcSurf);
-            if (frame < 0) {
-               frame = 0;
-            }
-            else if (frame >= frameCount) {
-               frame = frameCount - 1;
-            }
-         }
-
-         srcPixels += frame * Image_width(srcSurf);
-         dstPixels += dstY * dstPitch + dstX;
-         dstRow = dstPitch - w;
-
-         for (row = 0; row < h; row++, dstPixels += dstRow) {
-            int32 sourceY = (int32)((srcY + row) / scaleH);
-            PixelConv *srcRow = (PixelConv*)(srcPixels + sourceY * srcPitch);
-            for (col = 0; col < w; col++, dstPixels++) {
-               int32 sourceX = (int32)((srcX + col) / scaleW);
-               PixelConv *ps = srcRow + sourceX;
-               PixelConv *pt = (PixelConv*)dstPixels;
-               int32 a = ps->a * alphaMask;
-               a = (a + 1 + (a >> 8)) >> 8; // alphaMask * a / 255
-               if (a == 0xFF) {
-                  pt->pixel = ps->pixel;
-               }
-               else if (a != 0) {
-                  int32 ma = 0xFF - a;
-                  int32 r = (a * ps->r + ma * pt->r);
-                  int32 g = (a * ps->g + ma * pt->g);
-                  int32 b = (a * ps->b + ma * pt->b);
-                  pt->r = (r + 1 + (r >> 8)) >> 8;
-                  pt->g = (g + 1 + (g >> 8)) >> 8;
-                  pt->b = (b + 1 + (b >> 8)) >> 8;
-               }
-            }
-         }
-
-         markDirty(currentContext, dstSurf, dstX, dstY, w, h);
-         return;
-      }
-
-      int32 id = Image_textureId(srcSurf);
-      if (Image_changed(srcSurf) || id == -1) {
-         TCObject pixelsObj = frameCount > 1 ? Image_pixelsOfAllFrames(srcSurf) : Image_pixels(srcSurf);
-         Pixel* pixels = (Pixel*)ARRAYOBJ_START(pixelsObj);
-         int32 width = frameCount > 1 ? Image_widthOfAllFrames(srcSurf) : Image_width(srcSurf);
-         int32 height = Image_height(srcSurf);
-         Image_textureId(srcSurf) = skia_makeBitmap(id, pixels, width, height);
-         Image_changed(srcSurf) = false;
-      }
-
-      if (doClip) {
-         skia_setClip(Get_Clip(dstSurf));
-         clipSet = true;
-      }
-      if (frameCount > 1) {
-         frame = Image_currentFrame(srcSurf);
-         if (frame < 0) {
-            frame = 0;
-         }
-         else if (frame >= frameCount) {
-            frame = frameCount - 1;
-         }
-      }
-
-      skia_drawSurface(0, Image_textureId(srcSurf),
-         (float)(srcX / scaleW + frame * Image_width(srcSurf)),
-         (float)(srcY / scaleH),
-         (float)((srcX + w) / scaleW + frame * Image_width(srcSurf)),
-         (float)((srcY + h) / scaleH),
-         (float)dstX, (float)dstY, (float)(dstX + w), (float)(dstY + h),
-         Image_alphaMask(srcSurf));
-      if (clipSet) {
-         skia_restoreClip();
-      }
-   }
-   else {
-      LOGD("Trying to draw a control surface into some other surface");
-   }
-
-   markDirty(currentContext, dstSurf, dstX, dstY, w, h);
-}
-#endif
-
-//   Device specific routine.
-//   Gets the color value of the pixel, using the current translation
-//   Returns -1 if error (out of clip bounds)
-static int32 getPixel(TCObject g, int32 x, int32 y)
-{
-   int32 ret = -1;
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   if (Graphics_clipX1(g) <= x && x < Graphics_clipX2(g) && Graphics_clipY1(g) <= y && y < Graphics_clipY2(g))
-   {
-      PixelConv p;
-#ifdef SKIA_H
-        return skia_getPixel(0, x, y);
-#else
-#ifdef __gl2_h_
-      if (Graphics_useOpenGL(g))
-         return glGetPixel(x,y);
-      else
-#endif
-         p.pixel = getGraphicsPixels(g)[y * Graphics_pitch(g) + x];
-      ret = (p.r << 16) | (p.g << 8) | p.b;
-#endif
-   }
-   return ret;
-}
-
-#ifndef SKIA_H
-static PixelConv getPixelConv(TCObject g, int32 x, int32 y)
-{
-   PixelConv p;
-   p.pixel = -1;
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   if (Graphics_clipX1(g) <= x && x < Graphics_clipX2(g) && Graphics_clipY1(g) <= y && y < Graphics_clipY2(g))
-   {
-#ifdef __gl2_h_
-      if (Graphics_useOpenGL(g))
-         p.pixel = glGetPixel(x,y);
-      else
-#endif
-         p.pixel = getGraphicsPixels(g)[y * Graphics_pitch(g) + x];
-   }
-   return p;
-}
-#endif
-
-// Device specific routine.
-// Sets the pixel to the given color, translating and clipping
-#ifndef SKIA_H
-static void setPixel(Context currentContext, TCObject g, int32 x, int32 y, Pixel pixel)
-{
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   if (Graphics_clipX1(g) <= x && x < Graphics_clipX2(g) && Graphics_clipY1(g) <= y && y < Graphics_clipY2(g))
-   {
-#ifdef __gl2_h_
-      if (Graphics_useOpenGL(g))
-      {
-         glDrawPixel(x,y,pixel,255);
-         if (Graphics_isImageSurface(g))
-            Image_changed(Graphics_surface(g)) = true;
-         else
-            currentContext->fullDirty = true;
-      }
-      else
-#endif
-      {
-         getGraphicsPixels(g)[y * Graphics_pitch(g) + x] = pixel;
-         if (!currentContext->fullDirty && !Graphics_isImageSurface(g)) markScreenDirty(currentContext, x, y, 1, 1);
-      }
-   }
-}
-#else
-static void setPixel(Context currentContext, TCObject g, int32 x, int32 y, Pixel pixel)
-{
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   skia_setClip(Get_Clip(g));
-   skia_setPixel(0, x, y, pixel | Graphics_alpha(g));
-   skia_restoreClip();
-
-   markDirty(currentContext, g, x, y, 1, 1);
-}
-#endif
+#include "GraphicsPrimitivesSkia_c.h"
 
 #ifndef SKIA_H
 
@@ -1127,1209 +698,9 @@ static void fadeScreen(Context currentContext, int32 amount) {
 }
 #endif
 
-#ifndef SKIA_H
-#define INTERP(j,f,shift) (j + (((f - j) * transparency) >> shift)) & 0xFF
+#include "GraphicsPrimitivesText_c.h"
 
-static uint8 _ands8[8] = {0x80,0x40,0x20,0x10,0x08,0x04,0x02,0x01};
-uint8* getResizedCharPixels(Context currentContext, UserFont uf, JChar ch, int32 w, int32 h);
-
-static void drawText(Context currentContext, TCObject g, JCharP text, int32 chrCount, int32 x0, int32 y0, Pixel foreColor, int32 justifyWidth)
-{
-   TCObject fontObj = Graphics_font(g);
-   int32 startBit, currentBit, incY, y1, r, rmax, istart;
-   uint8 *bitmapTable, *ands, *current, *start;
-   uint16* bitIndexTable;
-   int32 rowWIB, offset, xMin, xMax, yMin, yMax, x, y, yDif, width, width0, height, spaceW = 0, k, clipX1,clipX2,clipY1,clipY2, pitch;
-   Pixel transparency, *row0, *row;
-   PixelConv *i;
-   bool isNibbleStartingLow, isLowNibble, isClipped;
-   int aaType;
-   JChar ch, first, last;
-   UserFont uf = null;
-   PixelConv fc;
-   int32 extraPixelsPerChar = 0, extraPixelsRemaining = -1, rem;
-   uint8 *ands8 = _ands8;
-   int32 fcR, fcG, fcB;
-#ifdef __gl2_h_
-   int32 charXY[2];
-   float *xya;
-#endif
-   int32 diffW;
-   bool isVert = Graphics_isVerticalText(g);
-   bool isGL = Graphics_useOpenGL(g);
-
-   if (!text || chrCount == 0 || fontObj == null) return;
-
-   fc.pixel = foreColor;
-   fcR = fc.r;
-   fcG = fc.g;
-   fcB = fc.b;
-
-   uf = loadUserFontFromFontObj(currentContext, fontObj, ' ');
-   if (uf == null) return;
-   diffW = uf->ubase && uf->isDefaultFont;
-   rowWIB = uf->rowWidthInBytes;
-   bitIndexTable = uf->bitIndexTable;
-   bitmapTable = uf->bitmapTable;
-   first = uf->fontP.firstChar;
-   last = uf->fontP.lastChar;
-
-   aaType = uf->fontP.antialiased;
-   height = uf->fontP.maxHeight;
-   incY = height + justifyWidth;
-
-   x0 += Graphics_transX(g);
-   y0 += Graphics_transY(g);
-
-   if (justifyWidth > 0)
-   {
-      while (chrCount > 0 && text[chrCount - 1] <= (JChar)' ')
-         chrCount--;
-      if (chrCount == 0) return;
-      rem = justifyWidth - getJCharPWidth(currentContext, fontObj, text, chrCount);
-      if (rem > 0)
-      {
-         extraPixelsPerChar = rem / chrCount;
-         extraPixelsRemaining = rem % chrCount;
-      }
-   }
-   clipX1 = Graphics_clipX1(g);
-   clipX2 = Graphics_clipX2(g);
-   clipY1 = Graphics_clipY1(g);
-   clipY2 = Graphics_clipY2(g);
-
-   xMax = xMin = (x0 < clipX1) ? clipX1 : x0;
-   yMax = y0 + (isVert ? chrCount * incY : height);
-   yMin = (y0 < clipY1) ? clipY1 : y0;
-   if (yMax >= clipY2)
-      yMax = clipY2;
-   if (getGraphicsPixels(g) == null)
-      return;
-   row0 = getGraphicsPixels(g) + yMin * Graphics_pitch(g);
-   yDif = yMin - y0;
-   y = y0;
-
-   pitch = Graphics_pitch(g);
-   for (k = 0; k < chrCount; k++) // guich@402
-   {
-      ch = *text++;
-      if (ch <= ' ' || ch == 160)
-      {
-         if (ch == ' ' || ch == '\t' || ch == 160)
-         {
-            if (isVert)
-               y += ch == '\t' ? incY * *tabSizeField : incY;
-            else
-            {
-               x0 += getJCharWidth(currentContext, fontObj, ch)+extraPixelsPerChar;
-               if (k <= extraPixelsRemaining)
-                  x0++;
-            }
-         }
-         continue;
-      }
-      if (uf == null || ch < first || ch > last)
-      {
-         uf = loadUserFontFromFontObj(currentContext, fontObj, ch);
-         if (uf == null || ch < uf->fontP.firstChar || ch > uf->fontP.lastChar) // invalid char - guich@tc122_23: must also check the font's range
-         {
-            x0 += spaceW ? spaceW : (spaceW=getJCharWidth(currentContext, fontObj, ' ')) + extraPixelsPerChar;
-            if (k <= extraPixelsRemaining)
-               x0++;
-            continue;
-         }
-         rowWIB = uf->rowWidthInBytes;
-         bitIndexTable = uf->bitIndexTable;
-         bitmapTable = uf->bitmapTable;
-         first = uf->fontP.firstChar;
-         last = uf->fontP.lastChar;
-      }
-#ifdef __gl2_h_
-      if (!checkGLfloatBuffer(currentContext, uf->fontP.maxHeight * uf->fontP.maxWidth))
-         return;
-#endif
-      // valid char, get its start
-      offset = bitIndexTable[ch];
-      width0 = width = bitIndexTable[ch+1] - offset - diffW;
-      isClipped = false;
-
-      if (uf->ubase != null) width = width * height / uf->ubase->fontP.maxHeight;
-      
-      if ((xMax = x0 + width) > clipX2)
-      {
-         isClipped = true;
-         xMax = clipX2;   
-      }
-      y1 = y; r=0;
-      istart = 0;
-      if (!isVert)
-      {
-         if (y0 < yMin) // guich@tc100b4_1: skip rows before yMin
-            istart += yMin-y0;
-         y = yMin;
-      }
-      else
-      if (y < yMin)
-      {
-         r += yMin-y;
-         istart += yMin-y; // guich@tc100b4_1: skip rows before yMin
-         y = yMin;
-      }
-      row0 = getGraphicsPixels(g) + y * Graphics_pitch(g);
-      rmax = (y+height > yMax) ? yMax - y : height;
-      isClipped |= x0 < clipX1 || istart != 0 || rmax != height;
-
-      switch (aaType)
-      {
-         case AA_NO:
-         {
-            start     = bitmapTable + (offset >> 3) + rowWIB * istart;
-            startBit  = offset & 7;
-
-            // draws the char, a row at a time
-   #ifdef __gl2_h_
-            if (isGL)
-            {
-               int32 nn=0;
-               
-               xya = glXYA;
-               for (; r < rmax; start+=rowWIB, r++,row += pitch,y++)    // draw each row
-               {
-                  current = start;
-                  ands = ands8 + (currentBit = startBit);
-                  for (x=x0; x < xMax; x++)
-                  {
-                     if ((*current & *ands++) != 0 && x >= xMin)
-                     {
-                        *xya++ = (float)x;
-                        *xya++ = (float)y;
-                        *xya++ = 1;
-                        nn++;
-                     }
-                     if (++currentBit == 8)   // finished this uint8?
-                     {
-                        currentBit = 0;       // reset counter
-                        ands = ands8;         // reset test bit pointer
-                        ++current;            // inc current uint8
-                     }
-                  }
-               }
-               if (nn > 0) // flush vertices buffer
-                  glDrawPixels(nn,foreColor);
-            }
-            else
-   #endif
-            for (row=row0; r < rmax; start+=rowWIB, r++,row += pitch)    // draw each row
-            {
-               current = start;
-               ands = ands8 + (currentBit = startBit);
-               for (x=x0; x < xMax; x++)
-               {
-                  if ((*current & *ands++) != 0 && x >= xMin)
-                     row[x] = foreColor;
-                  if (++currentBit == 8)   // finished this uint8?
-                  {
-                     currentBit = 0;       // reset counter
-                     ands = ands8;         // reset test bit pointer
-                     ++current;            // inc current uint8
-                  }
-               }
-            }
-            break;
-         }
-         case AA_4BPP:
-         {
-            start = bitmapTable + (offset >> 1) + rowWIB * istart;
-            isNibbleStartingLow = (offset & 1) == 1;
-            // draws the char, a row at a time
-   #ifdef __gl2_h_
-            if (isGL)
-            {
-               int32 nn=0;
-               
-               xya = glXYA;
-               for (; r < rmax; start+=rowWIB, r++,y++)    // draw each row
-               {
-                  current = start;
-                  isLowNibble = isNibbleStartingLow;
-                  for (x=x0; x < xMax; x++)
-                  {
-                     transparency = isLowNibble ? (*current++ & 0xF) : ((*current >> 4) & 0xF);
-                     isLowNibble = !isLowNibble;
-                     if (transparency == 0 || x < xMin)
-                        continue;
-
-                     // alpha
-                     // vertices
-                     *xya++ = (float)x;
-                     *xya++ = (float)y;
-                     *xya++ = ftransp[transparency];
-                     nn++;
-                  }
-               }
-               if (nn > 0) // flush vertices buffer
-                  glDrawPixels(nn,foreColor);
-            }
-            else
-   #endif
-               for (row=row0; r < rmax; start+=rowWIB, r++,row += pitch)    // draw each row
-               {
-                  current = start;
-                  isLowNibble = isNibbleStartingLow;
-                  i = (PixelConv*)&row[x0];
-                  for (x=x0; x < xMax; x++,i++)
-                  {
-                     transparency = isLowNibble ? (*current++ & 0xF) : ((*current >> 4) & 0xF);
-                     isLowNibble = !isLowNibble;
-                     if (transparency == 0 || x < xMin)
-                        continue;
-                     if (transparency == 0xF)
-                        i->pixel = foreColor;
-                     else
-                     {
-                        i->r = INTERP(i->r, fcR, 4);
-                        i->g = INTERP(i->g, fcG, 4);
-                        i->b = INTERP(i->b, fcB, 4);
-                     }
-                  }
-               }
-         }
-         break;
-         case AA_8BPP: // textured font files
-         {
-            // draws the char, a row at a time
-   #ifdef __gl2_h_
-            if (isGL)
-            {       
-               if (!isClipped && getCharPosInTexture(currentContext, uf->ubase, ch, charXY))
-/*text*/          glDrawTexture(uf->ubase->textureId, 
-                               charXY[0], charXY[1], width0, uf->ubase->fontP.maxHeight, // source char position
-                               x0, y, width, height,                                     // target bitmap position
-                               uf->ubase->maxW, uf->ubase->maxH, &fc, 255);              // total bitmap size
-               else
-               {                         
-                  uint8* alpha = getResizedCharPixels(currentContext, uf->ubase, ch, width+diffW, height);
-                  if (alpha)
-                  {                             
-                     int32 nn=0;
-                     rowWIB = width+diffW;
-                     start = alpha + istart * rowWIB;
-                     xya = glXYA;
-                     for (; r < rmax; start+=rowWIB, r++,y++)    // draw each row
-                     {
-                        current = start;
-                        for (x=x0; x < xMax; x++)
-                        {
-                           transparency = *current++;
-                           if (transparency == 0 || x < xMin)
-                              continue;
-   
-                           // alpha
-                           // vertices
-                           *xya++ = (float)x;
-                           *xya++ = (float)y;
-                           *xya++ = f255[transparency];
-                           nn++;
-                        }
-                     }
-                     if (nn > 0) // flush vertices buffer
-                        glDrawPixels(nn,foreColor);
-                  }
-               }
-            }
-            else
-   #endif // case 2
-            {
-               uint8* alpha = getResizedCharPixels(currentContext, uf->ubase, ch, width+diffW, height);
-               if (alpha)
-               {                             
-                  rowWIB = width+diffW;
-                  start = alpha + istart * rowWIB;
-                  for (row=row0; r < rmax; start+=rowWIB, r++,row += pitch)    // draw each row
-                  {
-                     current = start;
-                     i = (PixelConv*)&row[x0];
-                     for (x=x0; x < xMax; x++,i++)
-                     {
-                        transparency = *current++;
-                        if (transparency == 0 || x < xMin)
-                           continue;
-                        if (transparency == 0xFF)
-                           i->pixel = foreColor;
-                        else
-                        {
-                           i->r = INTERP(i->r, fcR, 8);
-                           i->g = INTERP(i->g, fcG, 8);
-                           i->b = INTERP(i->b, fcB, 8);
-                        }
-                     }
-                  }
-               }
-            }
-         }
-      }
-      if (isVert)
-      {
-         y = y1 + incY;
-         if (y >= yMax)
-            break;
-      }
-      else
-      {
-         if (xMax >= clipX2)
-         {
-            xMax = clipX2;
-            break;
-         }
-         x0 = xMax; // next character
-         x0 += extraPixelsPerChar;
-         if (k <= extraPixelsRemaining)
-            x0++;
-      }
-   }
-#ifndef __gl2_h_
-   if (!currentContext->fullDirty && !Graphics_isImageSurface(g)) markScreenDirty(currentContext, xMin, yMin, (xMax - xMin), (yMax - yMin));
-#else
-   if (Graphics_isImageSurface(g))
-      Image_changed(Graphics_surface(g)) = true;
-   else
-      currentContext->fullDirty = true;
-#endif
-}
-#else
-static void drawText(Context currentContext, TCObject g, JCharP text, int32 chrCount, int32 x, int32 y, Pixel foreColor, int32 justifyWidth)
-{
-   TCObject fontObj = Graphics_font(g);
-   int32 fontSize = (int)(Font_size(fontObj) * (*tcSettings.screenDensityPtr));
-   int32 typefaceIndex = Font_skiaIndex(fontObj);
-
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   skia_setClip(Get_Clip(g));
-   skia_drawText(0, text, chrCount * sizeof(JChar), x, y + fontSize, foreColor | Graphics_alpha(g), justifyWidth, fontSize, typefaceIndex);
-   skia_restoreClip();
-
-   markDirty(currentContext, g, x, y, skia_stringWidth(text, chrCount * sizeof(JChar), typefaceIndex, fontSize), fontSize);
-}
-#endif
-
-static SurfaceType getSurfaceType(Context currentContext, TCObject surface)
-{
-   // cache class pointers for performance
-   return (surface != NULL && areClassesCompatible(currentContext, OBJ_CLASS(surface), "totalcross.ui.image.Image") == 1) == COMPATIBLE ? SURF_IMAGE : SURF_CONTROL;
-}
-
-#ifndef SKIA_H
-////////////////////////////////////////////////////////////////////////////
-static void quadPixel(Context currentContext, TCObject g, int32 xc, int32 yc, int32 x, int32 y, Pixel c)
-{
-   // draw 4 points using symetry
-   setPixel(currentContext, g,xc + x, yc + y, c);
-   setPixel(currentContext, g,xc + x, yc - y, c);
-   setPixel(currentContext, g,xc - x, yc + y, c);
-   setPixel(currentContext, g,xc - x, yc - y, c);
-}
-
-static void quadLine(Context currentContext, TCObject g, int32 xc, int32 yc, int32 x, int32 y, Pixel c)
-{
-   int32 w = x+x+1; // plus 1 for the drawHLine (draws to width-1)
-   // draw 2 lines using symetry
-   drawHLine(currentContext, g,xc - x, yc - y, w, c, c);
-   drawHLine(currentContext, g,xc - x, yc + y, w, c, c);
-}
-
-// draws an ellipse incrementally
-static void ellipseDrawAndFill(Context currentContext, TCObject g, int32 xc, int32 yc, int32 rx, int32 ry, Pixel pc1, Pixel pc2, bool fill, bool gradient)
-{
-   int32 numSteps=0, startRed=0, startGreen=0, startBlue=0, endRed=0, endGreen=0, endBlue=0, redInc=0, greenInc=0, blueInc=0, red=0, green=0, blue=0;
-   PixelConv c,c1,c2;
-   // intermediate terms to speed up loop
-   int64 t1 = (int64)rx*(int64)rx, t2 = t1<<1, t3 = t2<<1;
-   int64 t4 = (int64)ry*(int64)ry, t5 = t4<<1, t6 = t5<<1;
-   int64 t7 = (int64)rx*t5, t8 = t7<<1, t9 = 0L;
-   int64 d1 = t2 - t7 + (t4>>1);    // error terms
-   int64 d2 = (t1>>1) - t8 + t5;
-   int32 x = rx;      // ellipse points
-   int32 y = 0;       // ellipse points
-   if (rx < 0 || ry < 0) // guich@501_13
-      return;
-   c1.pixel = pc1;
-   c2.pixel = pc2;
-
-   if (gradient)
-   {
-      numSteps = ry + ry; // guich@tc110_11: support horizontal gradient
-      startRed   = c1.r;
-      startGreen = c1.g;
-      startBlue = c1.b;
-      endRed = c2.r;
-      endGreen = c2.g;
-      endBlue = c2.b;
-      redInc = ((endRed - startRed) << 16) / numSteps;
-      greenInc = ((endGreen - startGreen) << 16) / numSteps;
-      blueInc = ((endBlue - startBlue) << 16) / numSteps;
-      red = startRed << 16;
-      green = startGreen << 16;
-      blue = startBlue << 16;
-   }
-   else c.pixel = c1.pixel;
-
-   while (d2 < 0)          // til slope = -1
-   {
-      if (gradient)
-      {
-         c.r = (red >> 16) & 0xFF;
-         c.g = (green >> 16) & 0xFF;
-         c.b = (blue >> 16) & 0xFF;
-         red += redInc;
-         green += greenInc;
-         blue += blueInc;
-      }
-      if (fill)
-         quadLine(currentContext, g,xc,yc,x,y,c.pixel);
-      else
-         quadPixel(currentContext, g,xc,yc,x,y,c.pixel);
-      y++;          // always move up here
-      t9 += t3;
-      if (d1 < 0)   // move straight up
-      {
-         d1 += t9 + t2;
-         d2 += t9;
-      }
-      else        // move up and left
-      {
-         --x;
-         t8 -= t6;
-         d1 += t9 + t2 - t8;
-         d2 += t9 + t5 - t8;
-      }
-   }
-
-   do             // rest of top right quadrant
-   {
-      if (gradient)
-      {
-         c.r = (red >> 16) & 0xFF;
-         c.g = (green >> 16) & 0xFF;
-         c.b = (blue >> 16) & 0xFF;
-         red += redInc;
-         green += greenInc;
-         blue += blueInc;
-      }
-      // draw 4 points using symmetry
-      if (fill)
-         quadLine(currentContext, g,xc,yc,x,y,c.pixel);
-      else
-         quadPixel(currentContext, g,xc,yc,x,y,c.pixel);
-      --x;        // always move left here
-      t8 -= t6;
-      if (d2 < 0)  // move up and left
-      {
-         ++y;
-         t9 += t3;
-         d2 += t9 + t5 - t8;
-      }
-      else d2 += t5 - t8; // move straight left
-   } while (x >= 0);
-}
-#else
-static void ellipseDrawAndFill(Context currentContext, TCObject g, int32 xc, int32 yc, int32 rx, int32 ry, Pixel pc1, Pixel pc2, bool fill, bool gradient)
-{
-   xc += Graphics_transX(g);
-   yc += Graphics_transY(g);
-   skia_setClip(Get_Clip(g));
-   skia_ellipseDrawAndFill(0, xc, yc, rx, ry, pc1 | Graphics_alpha(g), pc2 | Graphics_alpha(g), fill, gradient);
-   skia_restoreClip();
-
-   markDirty(currentContext, g, xc - rx, yc + ry, rx * 2, ry * 2);
-}
-#endif
-
-#ifndef SKIA_H
-////////////////////////////////////////////////////////////////////////////
-// Generalized Polygon Fill
-static void qsortInts(int32 *items, int32 first, int32 last)
-{
-   int32 low = first;
-   int32 high = last, mid;
-   if (first >= last)
-      return;
-   mid = items[(first+last) >> 1];
-   while (true)
-   {
-      while (high >= low && items[low] < mid) // guich@566_25: added "high > low" here and below - guich@568_5: changed to >=
-         low++;
-      while (high >= low && items[high] > mid)
-         high--;
-      if (low <= high)
-      {
-         int32 temp = items[low];
-         items[low++] = items[high];
-         items[high--] = temp;
-      }
-      else break;
-   }
-   if (first < high)
-      qsortInts(items, first,high);
-   if (low < last)
-      qsortInts(items, low,last);
-}
-
-static TCObject growIntArray(Context currentContext, TCObject oldArrayObj, int32 newLen) // must unlock the returned obj
-{
-   TCObject newArrayObj = createArrayObject(currentContext, INT_ARRAY, newLen);
-   int32 *newArray,*oldArray, oldLen;
-   if (newArrayObj != null)
-   {
-      newArray = (int32*)ARRAYOBJ_START(newArrayObj);
-      oldArray = (int32*)ARRAYOBJ_START(oldArrayObj);
-      oldLen = ARRAYOBJ_LEN(oldArrayObj);
-      xmemmove(newArray, oldArray, oldLen * 4);
-   }
-   return newArrayObj;
-}
-
-static void fillPolygon(Context currentContext, TCObject g, int32 *xPoints1, int32 *yPoints1, int32 nPoints1, int32 *xPoints2, int32 *yPoints2, int32 nPoints2, int32 tx, int32 ty, Pixel c1, Pixel c2, bool gradient, bool isPie)
-{
-   int32 x1, y1, x2, y2,y,n=0,temp, i,j, miny, maxy, a, numSteps=0, startRed=0, startGreen=0, startBlue=0, endRed=0, endGreen=0, endBlue=0, redInc=0, greenInc=0, blueInc=0, red=0, green=0, blue=0;
-   int32 *yp;
-   int32 *axPoints[2], *ayPoints[2], anPoints[2];
-   TCObject *intsObj = &Graphics_ints(g);
-   int32 *ints = *intsObj ? (int32*)ARRAYOBJ_START(*intsObj) : null;
-   PixelConv c;
-
-   if (!xPoints1 || !yPoints1 || nPoints1 < 2)
-      return;
-
-#if defined __gl2_h_
-   if (!gradient && (nPoints1 == 0 || isConvexAndInsideClip(g, tx, ty, xPoints1, yPoints1, nPoints1, isPie)) && (nPoints2 == 0 || isConvexAndInsideClip(g, tx, ty, xPoints2, yPoints2, nPoints2, isPie)) && Graphics_useOpenGL(g)) // opengl doesnt fills non-convex polygons well
-   {
-      if (nPoints1 > 0)
-         glDrawLines(currentContext, g, xPoints1, yPoints1, nPoints1, tx + Graphics_transX(g), ty + Graphics_transY(g), c1, true);
-      if (nPoints2 > 0)
-         glDrawLines(currentContext, g, xPoints2, yPoints2, nPoints2, tx + Graphics_transX(g), ty + Graphics_transY(g), c1, true);
-      return;
-   }
-#endif
-
-   axPoints[0] = xPoints1; ayPoints[0] = yPoints1; anPoints[0] = nPoints1;
-   axPoints[1] = xPoints2; ayPoints[1] = yPoints2; anPoints[1] = nPoints2;
-
-   yp = yPoints1;
-   miny = maxy = *yp++;
-   for (i = nPoints1; --i > 0; yp++)
-   {
-      if (*yp < miny) miny = *yp;
-      if (*yp > maxy) maxy = *yp;
-   }
-   yp = yPoints2;
-   for (i = nPoints2; --i >= 0; yp++)
-   {
-      if (*yp < miny) miny = *yp;
-      if (*yp > maxy) maxy = *yp;
-   }
-   miny += ty;
-   maxy += ty;
-
-   if (ints == null)
-   {
-      *intsObj = createArrayObject(currentContext, INT_ARRAY, 2); // 2 is the most used length
-      if (*intsObj == null)
-         return;
-      setObjectLock(*intsObj, UNLOCKED);
-      ints = (int32*)ARRAYOBJ_START(*intsObj);
-   }
-   if (gradient)
-   {
-      numSteps = maxy - miny; // guich@tc110_11: support horizontal gradient
-      if (numSteps == 0) numSteps = 1; // guich@tc115_86: prevent divide by 0
-      c.pixel = c1;
-      startRed   = c.r;
-      startGreen = c.g;
-      startBlue  = c.b;
-      c.pixel = c2;
-      endRed   = c.r;
-      endGreen = c.g;
-      endBlue  = c.b;
-      redInc = ((endRed - startRed) << 16) / numSteps;
-      greenInc = ((endGreen - startGreen) << 16) / numSteps;
-      blueInc = ((endBlue - startBlue) << 16) / numSteps;
-      red = startRed << 16;
-      green = startGreen << 16;
-      blue = startBlue << 16;
-   }
-   else c.pixel = c1;
-   for (y = miny; y <= maxy; y++)
-   {
-      n = 0;
-      for (a = 0; a < 2; a++)
-      {
-         int32 nPoints = anPoints[a];
-         int32* xPoints = axPoints[a];
-         int32* yPoints = ayPoints[a];
-         j = nPoints-1;
-         for (i = 0; i < nPoints; j=i,i++)
-         {
-            y1 = yPoints[j]+ty;
-            y2 = yPoints[i]+ty;
-            if (y1 == y2)
-               continue;
-            if (y1 > y2) // invert
-            {
-               temp = y1;
-               y1 = y2;
-               y2 = temp;
-            }
-            // compute next x point
-            if ( (y1 <= y && y < y2) || (y == maxy && y1 < y && y <= y2) )
-            {
-               if (n == (int32)ARRAYOBJ_LEN(*intsObj)) // have to grow the ints array?
-               {
-                  TCObject newIntsObj = growIntArray(currentContext, *intsObj, n * 2);
-                  if (newIntsObj == null)
-                     return;
-                  *intsObj = newIntsObj;
-                  setObjectLock(*intsObj, UNLOCKED);
-                  ints = (int32*)ARRAYOBJ_START(*intsObj);
-               }
-               if (yPoints[j] < yPoints[i])
-               {
-                  x1 = xPoints[j]+tx;
-                  x2 = xPoints[i]+tx;
-               }
-               else
-               {
-                  x2 = xPoints[j]+tx;
-                  x1 = xPoints[i]+tx;
-               }
-               ints[n++] = (y - y1) * (x2 - x1) / (y2 - y1) + x1;
-            }
-         }
-      }
-      if (n >= 2)
-      {
-         if (gradient)
-         {
-            c.r = (red   >> 16) & 0xFF;
-            c.g = (green >> 16) & 0xFF;
-            c.b = (blue  >> 16) & 0xFF;
-            red += redInc;
-            green += greenInc;
-            blue += blueInc;
-         }
-         if (n == 2) // most of the times
-         {
-            if (ints[1] > ints[0])
-               drawHLine(currentContext, g,ints[0],y,ints[1]-ints[0],c.pixel,c.pixel);
-            else
-               drawHLine(currentContext, g,ints[1],y,ints[0]-ints[1],c.pixel,c.pixel);
-         }
-         else
-         {
-            qsortInts(ints, 0, n-1);
-            for (n>>=1, yp = ints; --n >= 0; yp+=2)
-               drawHLine(currentContext, g,yp[0],y,yp[1]-yp[0],c.pixel,c.pixel);
-         }
-      }
-   }
-}
-#else
-static void fillPolygon(Context currentContext, TCObject g, int32 *xPoints1, int32 *yPoints1, int32 nPoints1, int32 *xPoints2, int32 *yPoints2, int32 nPoints2, int32 tx, int32 ty, Pixel c1, Pixel c2, bool gradient, bool isPie)
-{
-   skia_setClip(Get_Clip(g));
-   skia_fillPolygon(0, xPoints1, yPoints1, nPoints1, Graphics_transX(g), Graphics_transY(g), c1 | Graphics_alpha(g), c2 | Graphics_alpha(g), gradient, isPie);
-   skia_restoreClip();
-
-   // to avoid computing the polygon's bounds, we mark dirty the current clip
-   markDirty(currentContext, g, Graphics_clipX1(g), Graphics_clipY1(g), Graphics_clipX2(g) - Graphics_clipX1(g), Graphics_clipY2(g) - Graphics_clipY1(g));
-}
-#endif
-
-////////////////////////////////////////////////////////////////////////////
-// draws a polygon. if the polygon is not closed, close it
-#ifndef SKIA_H
-static void drawPolygon(Context currentContext, TCObject g, int32 *xPoints1, int32 *yPoints1, int32 nPoints1, int32 *xPoints2, int32 *yPoints2, int32 nPoints2, int32 tx, int32 ty, Pixel pixel)
-{
-   if (xPoints1 && yPoints1 && nPoints1 >= 2)
-   {
-#if defined __gl2_h_
-      if (Graphics_useOpenGL(g) && (nPoints1 == 0 || isInsideClip(g, tx, ty, xPoints1, yPoints1, nPoints1)) && (nPoints2 == 0 || isInsideClip(g, tx, ty, xPoints2, yPoints2, nPoints2)))
-      {
-         if (nPoints1 > 0)
-            glDrawLines(currentContext, g, xPoints1, yPoints1, nPoints1, tx + Graphics_transX(g), ty + Graphics_transY(g), pixel, false);
-         if (nPoints2 > 0)
-            glDrawLines(currentContext, g, xPoints2, yPoints2, nPoints2, tx + Graphics_transX(g), ty + Graphics_transY(g), pixel, false);
-      } 
-      else
-#endif
-      {
-         int32 i;
-         for (i=1; i < nPoints1; i++)
-            drawLine(currentContext, g,tx + xPoints1[i-1], ty + yPoints1[i-1], tx + xPoints1[i], ty + yPoints1[i], pixel);
-         for (i=1; i < nPoints2; i++)
-            drawLine(currentContext, g,tx + xPoints2[i-1], ty + yPoints2[i-1], tx + xPoints2[i], ty + yPoints2[i], pixel);
-      }
-   }
-}
-#else
-static void drawPolygon(Context currentContext, TCObject g, int32 *xPoints1, int32 *yPoints1, int32 nPoints1, int32 *xPoints2, int32 *yPoints2, int32 nPoints2, int32 tx, int32 ty, Pixel pixel)
-{
-   skia_setClip(Get_Clip(g));
-   skia_drawPolygon(0, xPoints1, yPoints1, nPoints1, Graphics_transX(g), Graphics_transY(g), pixel | Graphics_alpha(g));
-   skia_restoreClip();
-
-   // to avoid computing the polygon's bounds, we mark dirty the current clip
-   markDirty(currentContext, g, Graphics_clipX1(g), Graphics_clipY1(g), Graphics_clipX2(g) - Graphics_clipX1(g), Graphics_clipY2(g) - Graphics_clipY1(g));
-}
-#endif
-////////////////////////////////////////////////////////////////////////////
-// draw an elliptical arc from startAngle to endAngle.
-// c is the fill color and c2 is the outline color
-// (if in fill mode - otherwise, c = outline color)
-#ifdef SKIA_H
-static void arcPiePointDrawAndFill(Context currentContext, TCObject g, int32 xc, int32 yc, int32 rx, int32 ry, double startAngle, double endAngle, Pixel c, Pixel c2, bool fill, bool pie, bool gradient)
-{
-   xc += Graphics_transX(g);
-   yc += Graphics_transY(g);
-   skia_setClip(Get_Clip(g));
-   skia_arcPiePointDrawAndFill(0, xc, yc, rx, ry, startAngle, endAngle, c | Graphics_alpha(g), c2 | Graphics_alpha(g), fill, pie, gradient);
-   skia_restoreClip();
-
-   markDirty(currentContext, g, xc - rx, yc + ry, rx * 2, ry * 2);
-}
-#else
-static void arcPiePointDrawAndFill(Context currentContext, TCObject g, int32 xc, int32 yc, int32 rx, int32 ry, double startAngle, double endAngle, Pixel c, Pixel c2, bool fill, bool pie, bool gradient)
-{
-   // this algorithm was created by Guilherme Campos Hazan
-   double ppd;
-   int32 startIndex,endIndex,index,i,nq,size=0,oldX1=0,oldY1=0,last,oldX2=0,oldY2=0;
-   bool sameR,startSetTo0 = true;
-   TCObject *xPointsObj = &Graphics_xPoints(g);
-   TCObject *yPointsObj = &Graphics_yPoints(g);
-   int32 *xPoints = *xPointsObj ? (int32*)ARRAYOBJ_START(*xPointsObj) : null;
-   int32 *yPoints = *yPointsObj ? (int32*)ARRAYOBJ_START(*yPointsObj) : null;
-   int32 clipFactor = Graphics_minX(g) * 1000000000 + Graphics_maxX(g) * 10000000 + Graphics_minY(g) * 100000 + Graphics_maxY(g);
-   bool sameClipFactor = Graphics_lastClipFactor(g) == clipFactor;
-
-   if (rx < 0 || ry < 0) // guich@501_13
-      return;
-   // make sure the values are -359 <= x <= 359
-   while (startAngle <= -360) startAngle += 360;
-   while (endAngle   <= -360) endAngle   += 360;
-   while (startAngle >   360) startAngle -= 360;
-   while (endAngle   >   360) endAngle   -= 360;
-
-   if (startAngle == endAngle) // guich@501_13
-      return;
-   if (startAngle > endAngle) // eg 235 to 45
-      startAngle -= 360; // set to -45 to 45 so we can handle it correctly
-   if (startAngle >= 0 && endAngle <= 0) // eg 135 to -135
-      endAngle += 360; // set to 135 to 225
-
-   // step 0: correct angle values
-   if (startAngle < 0.1 && endAngle > 359.9) // full circle? use the fastest routine instead
-   {
-      if (fill)
-         ellipseDrawAndFill(currentContext, g,xc, yc, rx, ry, c, c2, true, gradient);
-      ellipseDrawAndFill(currentContext, g,xc, yc, rx, ry, c, c, false, gradient);
-      return;
-   }
-
-   // step 0: if possible, use cached results
-   sameR = rx == Graphics_lastRX(g) && ry == Graphics_lastRY(g);
-   if (!sameClipFactor || !sameR)
-   {
-      // step 1: computes how many points the circle has (computes only 45 degrees and mirrors the rest)
-      // intermediate terms to speed up loop
-      int64 t1 = (int64)rx*(int64)rx, t2 = t1<<1, t3 = t2<<1;
-      int64 t4 = (int64)ry*(int64)ry, t5 = t4<<1, t6 = t5<<1;
-      int64 t7 = (int64)rx*t5, t8 = t7<<1, t9 = 0L;
-      int64 d1 = t2 - t7 + (t4>>1);    // error terms
-      int64 d2 = (t1>>1) - t8 + t5;
-      int32 x = rx;                 // ellipse points
-      int32 y = 0;                  // ellipse points
-
-      while (d2 < 0)              // til slope = -1
-      {
-         t9 += t3;
-         if (d1 < 0)             // move straight up
-         {
-            d1 += t9 + t2;
-            d2 += t9;
-         }
-         else                   // move up and left
-         {
-            --x;
-            t8 -= t6;
-            d1 += t9 + t2 - t8;
-            d2 += t9 + t5 - t8;
-         }
-         ++size;
-      }
-
-      do             // rest of top right quadrant
-      {
-         --x;         // always move left here
-         t8 -= t6;
-         if (d2 < 0)  // move up and left
-         {
-            t9 += t3;
-            d2 += t9 + t5 - t8;
-         }
-         else d2 += t5 - t8;  // move straight left
-         ++size;
-      } while (x >= 0);
-      nq = size;
-      size *= 4;
-      // step 2: computes how many points per degree
-      ppd = (double)size / 360.0f;
-      // step 3: create space in the buffer so it can save all the circle
-      size+=2;
-      if (pie) size++;
-      if (xPoints == null || ARRAYOBJ_LEN(*xPointsObj) != (uint32)size) // guich@tc304: changed < to != to fix a glytch when drawing two pies with different radius
-      {
-         *xPointsObj = createArrayObject(currentContext, INT_ARRAY, max32(3,size));
-         if (*xPointsObj == null)
-            return;
-         *yPointsObj = createArrayObject(currentContext, INT_ARRAY, max32(3,size));
-         if (*yPointsObj == null)
-         {
-            setObjectLock(*xPointsObj, UNLOCKED);
-            return;
-         }
-         setObjectLock(*xPointsObj, UNLOCKED);
-         setObjectLock(*yPointsObj, UNLOCKED);
-      }
-      xPoints = (int32*)ARRAYOBJ_START(*xPointsObj);
-      yPoints = (int32*)ARRAYOBJ_START(*yPointsObj);
-      if (pie) {xPoints++; yPoints++;} // make sure that startIndex-1 is at a valid pointer
-
-      // step 4: stores all the circle in the array. the odd arcs are drawn in reverse order
-      // intermediate terms to speed up loop
-      t2 = t1<<1;
-      t3 = t2<<1;
-      t8 = t7<<1;
-      t9 = 0;
-      d1 = t2 - t7 + (t4>>1); // error terms
-      d2 = (t1>>1) - t8 + t5;
-      x = rx;
-      i=0;
-      while (d2 < 0)          // til slope = -1
-      {
-         // save 4 points using symmetry
-         index = nq*0+i;      // 0/3
-         xPoints[index]=+x;
-         yPoints[index]=-y;
-
-         index = (nq<<1)-i-1;    // 1/3
-         xPoints[index]=-x;
-         yPoints[index]=-y;
-
-         index = (nq<<1)+i;      // 2/3
-         xPoints[index]=-x;
-         yPoints[index]=+y;
-
-         index = (nq<<2)-i-1;    // 3/3
-         xPoints[index]=+x;
-         yPoints[index]=+y;
-         i++;
-         y++;        // always move up here
-         t9 += t3;
-         if (d1 < 0)  // move straight up
-         {
-             d1 += t9 + t2;
-             d2 += t9;
-         }
-         else      // move up and left
-         {
-             x--;
-             t8 -= t6;
-             d1 += t9 + t2 - t8;
-             d2 += t9 + t5 - t8;
-         }
-      }
-
-      do             // rest of top right quadrant
-      {
-         // save 4 points using symmetry
-         index = nq*0+i;    // 0/3
-         xPoints[index]=+x;
-         yPoints[index]=-y;
-
-         index = (nq<<1)-i-1;  // 1/3
-         xPoints[index]=-x;
-         yPoints[index]=-y;
-
-         index = (nq<<1)+i;    // 2/3
-         xPoints[index]=-x;
-         yPoints[index]=+y;
-
-         index = (nq<<2)-i-1;  // 3/3
-         xPoints[index]=+x;
-         yPoints[index]=+y;
-
-         ++i;
-         --x;        // always move left here
-         t8 -= t6;
-         if (d2 < 0)  // move up and left
-         {
-            ++y;
-            t9 += t3;
-            d2 += t9 + t5 - t8;
-         }
-         else d2 += t5 - t8;   // move straight left
-      } while (x >= 0);
-      // save last arguments
-      //Graphics_lastXC(g)   = xc; no longer
-      //Graphics_lastYC(g)   = yc;  needed
-      Graphics_lastRX(g)   = rx;
-      Graphics_lastRY(g)   = ry;
-      Graphics_lastPPD(g)  = ppd;
-      Graphics_lastSize(g) = size;
-      Graphics_lastClipFactor(g) = clipFactor;
-   }
-   else
-   {
-      size = Graphics_lastSize(g);
-      ppd = Graphics_lastPPD(g);
-   }
-   // step 5: computes the start and end indexes that will become part of the arc
-   if (startAngle < 0)
-      startAngle += 360;
-   if (endAngle < 0)
-      endAngle += 360;
-   startIndex = (int32)(ppd * startAngle);
-   endIndex = (int32)(ppd * endAngle);
-
-   last = size-2;
-   if (endIndex >= last) // 360?
-      endIndex--;
-   // step 6: fill or draw the polygons
-   endIndex++;
-   if (pie)
-   {
-      // connect two lines from the center to the two edges of the arc
-      oldX1 = xPoints[endIndex];
-      oldY1 = yPoints[endIndex];
-      xPoints[endIndex] = yPoints[endIndex] = 0;
-      if (xPoints[startIndex] == 0 && yPoints[startIndex] == 0)
-         startSetTo0 = false;
-      else
-      {
-         startIndex--;
-         oldX2 = xPoints[startIndex];
-         oldY2 = yPoints[startIndex];
-         xPoints[startIndex] = yPoints[startIndex] = 0;
-      }
-      endIndex++;
-   }
-  
-   if (startIndex > endIndex) // drawing from angle -30 to +30 ? (startIndex = 781, endIndex = 73, size=854)
-   {
-      int p1 = last-startIndex;
-      if (fill)
-         fillPolygon(currentContext, g, xPoints+startIndex, yPoints+startIndex, p1, xPoints, yPoints, endIndex, xc,yc, gradient ? c : c2, c2, gradient, true); // lower half, upper half
-      if (!gradient) drawPolygon(currentContext, g, xPoints+startIndex, yPoints+startIndex, p1-1, xPoints+1, yPoints+1, endIndex-1, xc,yc, c);
-   }
-   else
-   {
-      int32 arc = pie ? 0 : 1;
-      if (fill)
-         fillPolygon(currentContext, g, xPoints+startIndex, yPoints+startIndex, endIndex-startIndex, 0,0,0, xc,yc, gradient ? c : c2, c2, gradient, true);   
-      if (!gradient) drawPolygon(currentContext, g, xPoints+startIndex+arc, yPoints+startIndex+arc, endIndex-startIndex-arc, 0,0,0, xc,yc, c);
-   }
-   if (pie)  // restore saved points
-   {
-      if (startSetTo0)
-      {
-         xPoints[startIndex] = oldX2;
-         yPoints[startIndex] = oldY2;
-      }
-      endIndex--;
-      xPoints[endIndex]   = oldX1;
-      yPoints[endIndex]   = oldY1;
-#ifdef ANDROID
-      if (!gradient && endAngle == 360) 
-         drawLine(currentContext,g, xc,yc, xc+xPoints[endIndex-1], yc+yPoints[endIndex-1], c);
-#endif         
-   }
-}
-#endif
-////////////////////////////////////////////////////////////////////////////
-#ifndef SKIA_H
-static void drawRoundRect(Context currentContext, TCObject g, int32 x, int32 y, int32 width, int32 height, int32 r, Pixel c)
-{
-   int32 x1, y1, x2, y2, dec, xx, yy;
-   int32 w, h;
-   r = min32(r,min32(width/2,height/2));
-   w = width - 2*r;
-   h = height - 2*r;
-   x1 = x+r;
-   y1 = y+r;
-   x2 = x+width-r-1;
-   y2 = y+height-r-1;
-   dec = 3-2*r;
-
-   drawHLine(currentContext, g,x+r, y, w, c, c); // top
-   drawHLine(currentContext, g,x+r, y+height-1, w, c, c); // bottom
-   drawVLine(currentContext, g,x, y+r, h, c, c); // left
-   drawVLine(currentContext, g,x+width-1, y+r, h, c, c); // right
-
-   // draw the round rectangles.
-   for (xx = 0, yy = r; xx <= yy; xx++)
-   {
-      setPixel(currentContext, g,x2+xx, y2+yy, c);
-      setPixel(currentContext, g,x2+xx, y1-yy, c);
-      setPixel(currentContext, g,x1-xx, y2+yy, c);
-      setPixel(currentContext, g,x1-xx, y1-yy, c);
-
-      setPixel(currentContext, g,x2+yy, y2+xx, c);
-      setPixel(currentContext, g,x2+yy, y1-xx, c);
-      setPixel(currentContext, g,x1-yy, y2+xx, c);
-      setPixel(currentContext, g,x1-yy, y1-xx, c);
-      if (dec >= 0)
-         dec += -4*(yy--)+4;
-      dec += 4*xx+6;
-   }
-}
-#else
-static void drawRoundRect(Context currentContext, TCObject g, int32 x, int32 y, int32 w, int32 h, int32 r, Pixel c)
-{
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   skia_setClip(Get_Clip(g));
-   skia_drawRoundRect(0, x, y, w, h, r, c | Graphics_alpha(g));
-   skia_restoreClip();
-
-   markDirty(currentContext, g, x, y, w, h);
-}
-#endif
-
-#ifndef SKIA_H
-////////////////////////////////////////////////////////////////////////////
-static void setPixelA(Context currentContext, TCObject g, int32 x, int32 y, PixelConv color, int32 alpha);
-
-////////////////////////////////////////////////////////////////////////////
-static void fillRoundRect(Context currentContext, TCObject g, int32 xx, int32 yy, int32 width, int32 height, int32 r, Pixel c)
-{
-   int32 px1,px2,py1,py2,xm,ym,x,y=0, i, x2, e2, err;
-   PixelConv color;
-   if (r > (width/2) || r > (height/2)) r = min32(width/2,height/2); // guich@200b4_6: correct bug that crashed the device.
-
-   x = -r;
-   err = 2 - 2 * r;
-   color.pixel = c;
-
-   px1 = xx+r;
-   py1 = yy+r;
-   px2 = xx+width-r-1;
-   py2 = yy+height-r-1;
-
-   height -= 2*r;
-   yy += r;
-   while (height--)
-      drawHLine(currentContext, g,xx, yy++, width, c, c);
-
-   r = 1 - err;
-   do
-   {
-      i = 255 - 255 * abs(err - 2 * (x + y) - 2) / r;
-
-      drawLine(currentContext, g, px1+x+1,py1-y,px2-x-1,py1-y,c);
-      drawLine(currentContext, g, px1+x+1,py2+y,px2-x-1,py2+y,c);
-
-      if (i < 256 && i > 0)
-      {
-         xm = px2; ym = py2; setPixelA(currentContext, g, xm - x, ym + y, color, i); // br
-         xm = px1; ym = py2; setPixelA(currentContext, g, xm - y, ym - x, color, i); // bl
-         xm = px1; ym = py1; setPixelA(currentContext, g, xm + x, ym - y, color, i); // tl
-         xm = px2; ym = py1; setPixelA(currentContext, g, xm + y, ym + x, color, i); // tr
-      }
-      e2 = err;
-      x2 = x;
-      if (err + y > 0)
-      {
-         i = 255 - 255 * (err - 2 * x - 1) / r;
-         if (i < 256 && i > 0)
-         {
-            xm = px2; ym = py2; setPixelA(currentContext, g, xm - x, ym + y + 1, color, i);
-            xm = px1; ym = py2; setPixelA(currentContext, g, xm - y - 1, ym - x, color, i);
-            xm = px1; ym = py1; setPixelA(currentContext, g, xm + x, ym - y - 1, color, i);
-            xm = px2; ym = py1; setPixelA(currentContext, g, xm + y + 1, ym + x, color, i);
-         }
-         err += ++x * 2 + 1;
-      }
-      if (e2 + x2 <= 0)
-      {
-         i = 255 - 255 * (2 * y + 3 - e2) / r;
-         if (i < 256 && i > 0)
-         {
-            xm = px2; ym = py2; setPixelA(currentContext, g, xm - x2 - 1, ym + y, color, i);
-            xm = px1; ym = py2; setPixelA(currentContext, g, xm - y, ym - x2 - 1, color, i);
-            xm = px1; ym = py1; setPixelA(currentContext, g, xm + x2 + 1, ym - y, color, i);
-            xm = px2; ym = py1; setPixelA(currentContext, g, xm + y, ym + x2 + 1, color, i);
-         }
-         err += ++y * 2 + 1;
-      }
-   } while (x < 0);
-}
-#else
-static void fillRoundRect(Context currentContext, TCObject g, int32 x, int32 y, int32 w, int32 h, int32 r, Pixel c)
-{
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   skia_setClip(Get_Clip(g));
-   skia_fillRoundRect(0, x, y, w, h, r, c | Graphics_alpha(g));
-   skia_restoreClip();
-
-   markDirty(currentContext, g, x, y, w, h);
-}
-#endif
-
-#if 1//ndef SKIA_H
-// Translates the given coords and returns the intersection between
-// the clip rect and the coords passed.
-// Returns: 1 if OK, 0 if the coords are outside the clip rect
-static bool translateAndClip(TCObject g, int32 *pX, int32 *pY, int32 *pWidth, int32 *pHeight)
-{
-   int32 x = *pX;
-   int32 y = *pY;
-   int32 w = *pWidth;
-   int32 h = *pHeight;
-   x += Graphics_transX(g);
-   y += Graphics_transY(g);
-   if (x < Graphics_clipX1(g))
-   {
-      if ((x+w) > Graphics_clipX2(g))
-         w = Graphics_clipX2(g) - Graphics_clipX1(g);
-      else
-         w -= Graphics_clipX1(g)-x;
-      x = Graphics_clipX1(g);
-   }
-   else
-   if ((x+w) > Graphics_clipX2(g))
-      w = Graphics_clipX2(g) - x;
-   if (y < Graphics_clipY1(g))
-   {
-      if ((y+h) > Graphics_clipY2(g))
-         h = Graphics_clipY2(g) - Graphics_clipY1(g);
-      else
-         h -= Graphics_clipY1(g)-y;
-      y = Graphics_clipY1(g);
-   }
-   else
-   if ((y+h) > Graphics_clipY2(g))
-      h = Graphics_clipY2(g) - y;
-
-   if (x < 0 || y < 0 || h <= 0 || w <= 0) return false; // guich@566_42: check the resulting w/h - guich@tc112_34: check also x and y
-
-   *pX      = x;
-   *pY      = y;
-   *pWidth  = w;
-   *pHeight = h;
-   return true;
-}
-#endif
+#include "GraphicsPrimitivesShapes_c.h"
 
 static void createGfxSurface(int32 w, int32 h, TCObject g, SurfaceType stype)
 {
@@ -3127,104 +1498,4 @@ static void dither(Context currentContext, TCObject g, int32 x0, int32 y0, int32
 }
 
 /////////////// Start of Device-dependant functions ///////////////
-static bool startupGraphics(int16 appTczAttr) // there are no threads running at this point
-{
-    return graphicsStartup(&screen, appTczAttr);
-}
-
-static bool createScreenSurface(Context currentContext, bool isScreenChange)
-{
-   bool ret = false;
-   if (screen.screenW <= 0 || screen.screenH <= 0)
-      return false;
-
-   if (graphicsCreateScreenSurface(&screen))
-   {
-      TCObject *screenObj;
-      screenObj = getStaticFieldObject(currentContext,loadClass(currentContext, "totalcross.ui.gfx.Graphics",false), "mainWindowPixels");
-#ifdef darwin // in darwin, the pixels buffer is pre-initialized and never changed
-      if (controlEnableUpdateScreenPtr == null)
-         controlEnableUpdateScreenPtr = getStaticFieldInt(loadClass(currentContext, "totalcross.ui.Control",false), "enableUpdateScreen");
-      ret = true;
-#else
-
-      if (isScreenChange)
-      {
-         screen.mainWindowPixels = *screenObj = null;
-         gc(currentContext); // let the gc collect the old screen object
-      }
-      else
-      {
-         controlEnableUpdateScreenPtr = getStaticFieldInt(loadClass(currentContext, "totalcross.ui.Control",false), "enableUpdateScreen");
-      }
-
-      *screenObj = screen.mainWindowPixels = createArrayObject(currentContext, INT_ARRAY, screen.screenW * screen.screenH);
-      setObjectLock(*screenObj, UNLOCKED);
-      ret = screen.mainWindowPixels != null && controlEnableUpdateScreenPtr != null;
-#endif
-   }
-   return ret;
-}
-
-void markWholeScreenDirty(Context currentContext)
-{
-   LOCKVAR(screen);
-   currentContext->dirtyX1 = currentContext->dirtyY1 = 0;
-   currentContext->dirtyX2 = screen.screenW;
-   currentContext->dirtyY2 = screen.screenH;
-   currentContext->fullDirty = true;
-   UNLOCKVAR(screen);
-}
-
-static bool checkScreenPixels()
-{
-   return screen.pixels != null;
-}
-
-void updateScreen(Context currentContext)
-{
-#ifdef ANDROID
-   if (appPaused) return;
-#endif
-   LOCKVAR(screen);
-   if (keepRunning && checkScreenPixels() && controlEnableUpdateScreenPtr && *controlEnableUpdateScreenPtr && (currentContext->fullDirty || (currentContext->dirtyX1 != screen.screenW && currentContext->dirtyX2 != 0 && currentContext->dirtyY1 != screen.screenH && currentContext->dirtyY2 != 0)))
-   {
-      if (updateScreenBits(currentContext)) // move the temporary buffer to the real screen
-      {
-#ifdef darwin
-         currentContext->dirtyX1 = screen.screenW;
-         currentContext->dirtyY1 = screen.screenH;
-         currentContext->dirtyX2 = currentContext->dirtyY2 = 0;
-         currentContext->fullDirty = false;
-         UNLOCKVAR(screen); // without this, a deadlock can occur in iOS if the user minimizes the application, since another thread can trigger a markScreenDirty
-#endif
-         graphicsUpdateScreen(currentContext, &screen);
-#ifdef darwin
-         LOCKVAR(screen);
-#else
-         currentContext->dirtyX1 = screen.screenW;
-         currentContext->dirtyY1 = screen.screenH;
-         currentContext->dirtyX2 = currentContext->dirtyY2 = 0;
-         currentContext->fullDirty = false;
-#endif
-      }
-   }
-   UNLOCKVAR(screen);
-#if defined ANDROID || defined darwin
-   if (desiredScreenShiftY != UNDEFINED_SHIFTY)
-   {
-      setShiftYgl(desiredScreenShiftY);
-      desiredScreenShiftY = UNDEFINED_SHIFTY;
-   }
-#endif
-}
-
-void graphicsDestroyPrimitives()
-{
-   xfree(lookupR);
-   xfree(lookupG);
-   xfree(lookupB);
-   xfree(lookupGray);
-   fontDestroy();
-}
-/////////////// End of Device-dependant functions ///
+#include "GraphicsPrimitivesScreen_c.h"
