+++ TotalCrossVM/src/nm/ui/skia/skia_surface.cpp
@@ -1,0 +1,1 @@
+// Copyright (C) 2026 Amalgam Solucoes em TI Ltda//// SPDX-License-Identifier: LGPL-2.1-only#include "skia_internal.h"static void releaseProc(void* addr, void* ) {    delete[] static_cast<int32*>(addr);}int skia_makeBitmap(int32 id, void *data, int32 w, int32 h) {    SKIA_TRACE()    const size_t count = (size_t)w * (size_t)h;    int32* const converted = new int32[count];    const Pixel* const src = reinterpret_cast<const Pixel*>(data);    for (size_t i = 0; i < count; i++) {        converted[i] = SWAP32(src[i]);    }    if (id < 0) { // must create a new bitmap        SkBitmap bitmap;        bitmap.installPixels(SkImageInfo::Make(w, h, kN32_SkColorType, kUnpremul_SkAlphaType),                             (void*)converted, sizeof(Pixel) * w, releaseProc, nullptr);        // TODO: Reevaluate this optimization; it was added for embedded Linux ARM, but forced scale draws broke on Android/iOS.#if USE_COMPUTE_OPAQUE        if (SkBitmap::ComputeIsOpaque(bitmap)) {            bitmap.setAlphaType(kOpaque_SkAlphaType);        }#endif        // TODO: Recheck the color type before re-enabling on Android/iOS; macOS did not reproduce the forced scale issue.#if USE_COLORTYPE_CONVERSION        if (bitmap.isOpaque()) {            void* dstPixels = operator new(w * h * canvas->imageInfo().bytesPerPixel());            SkImageInfo dstImageInfo =                SkImageInfo::Make(                    w,                    h,                    canvas->imageInfo().colorType(),                    canvas->imageInfo().alphaType()                );            bitmap.readPixels(dstImageInfo, dstPixels, dstImageInfo.minRowBytes(), 0, 0);            bitmap.installPixels(dstImageInfo, dstPixels, dstImageInfo.minRowBytes(), releaseProc, nullptr);        }#endif        id = textures.size();        textures.emplace_back(std::move(bitmap));    } else {        SkBitmap& bitmap = textures[id];        bitmap.installPixels(SkImageInfo::Make(w, h, kN32_SkColorType, kUnpremul_SkAlphaType), (void*)converted, sizeof(Pixel) * w, releaseProc, nullptr);    }    return id;}void skia_deleteBitmap(int32 id) {    SKIA_TRACE()    if (id >= 0) {        textures[id].reset();    }}void skia_setClip(int32 x1, int32 y1, int32 x2, int32 y2){    canvas->save();    canvas->clipRect(SkRect::MakeLTRB(x1, y1, x2, y2));}void skia_restoreClip(){    canvas->restore();}void skia_drawSurface(int32 skiaSurface, int32 id, float srcLeft, float srcTop, float srcRight, float srcBottom, float dstLeft, float dstTop, float dstRight, float dstBottom, int32 alphaMask){    SKIA_TRACE()	const auto& texture {textures[id]};    const SkRect srcRect = SkRect::MakeLTRB(srcLeft, srcTop, srcRight, srcBottom);    const SkRect dstRect = SkRect::MakeLTRB(dstLeft, dstTop, dstRight, dstBottom);    const bool fullSource = srcLeft == 0.0f && srcTop == 0.0f && srcRight == texture.width() && srcBottom == texture.height();    const bool sameSize = (srcRight - srcLeft) == (dstRight - dstLeft) && (srcBottom - srcTop) == (dstBottom - dstTop);#if USE_WRITE_PIXELS    /*        Fast drawing, can only be used to draw fully opaque images        without sampling (src and dst dimensions are the same).        Makes drawing JPEG and opaque PNGs over 10x faster.        TODO:            - add actual numbers to back this statement    */if (texture.isOpaque() && alphaMask == 255 && fullSource && sameSize) {    canvas->writePixels(        texture.info(),        texture.getPixels(),        texture.rowBytes(),        (int)dstLeft,        (int)dstTop    );} else#endif    {        alphaPaint.setAlpha(alphaMask);        alphaPaint.setFilterQuality(sameSize ? kNone_SkFilterQuality : kLow_SkFilterQuality);        canvas->drawBitmapRect(            texture,            srcRect,            dstRect,            &alphaPaint,            fullSource ? SkCanvas::kFast_SrcRectConstraint : SkCanvas::kStrict_SrcRectConstraint        );    }}// The getPixel call demands a 1-pixel readback from the GPU. Avoid it if possible.Pixel skia_getPixel(int32 skiaSurface, int32 x, int32 y){    SKIA_TRACE()    SkBitmap bitmap;    bitmap.allocPixels(SkImageInfo::MakeN32Premul(1, 1));    if (!canvas->readPixels(bitmap, x, y))    {        return -1;    }    Pixel pixel = bitmap.getAddr32(0, 0)[0];    return (((pixel >> 24) & 0xFF) << 24) | (((pixel & 0xFF) << 16) | (((pixel >> 8) & 0xFF) << 8) | ((pixel >> 16) & 0xFF));}void skia_setPixel(int32 skiaSurface, int32 x, int32 y, Pixel pixel){    SKIA_TRACE()    backPaint.setColor(pixel);    canvas->drawRect(SkRect::MakeXYWH(x, y, 1, 1), backPaint);}int skia_getsetRGB(int32 skiaSurface, void *pixels, int32 offset, int32 x, int32 y, int32 w, int32 h, bool isGet){    SKIA_TRACE()    // SkImageInfo info = SkImageInfo::MakeN32Premul(w, h);    SkBitmap bitmap;    // bitmap.installPixels(SkImageInfo::Make(w, h, kN32_SkColorType, kUnpremul_SkAlphaType), (void*)pixels, sizeof(Pixel) * w, nullptr, nullptr);    // SkBitmap bitmap;    bitmap.allocPixels(SkImageInfo::MakeN32Premul(w, h));    if (isGet) {        return canvas->readPixels(bitmap, x, y);    } else {        return canvas->writePixels(bitmap, x, y);    }    // Pixel pixel = bitmap.getAddr32(0, 0)[0];    // return (((pixel >> 24) & 0xFF) << 24) | (((pixel & 0xFF) << 16) | (((pixel >> 8) & 0xFF) << 8) | ((pixel >> 16) & 0xFF));    // if (isGet) {    //     return canvas->readPixels(info, pixels, info.minRowBytes(), x, y);    // } else {    //     return canvas->writePixels(info, pixels, info.minRowBytes(), x, y);    // }}
